memory: restrict access to clone frames
only code in the memory module should be able to clone frames, for
convenience when the situation is safe and makes sense. if we allow
outside functions to clone frames, we have no guarantees that a frame we
are acting on is unused without going through a ton of checking work.
this lets us rely on rusts ownership semantics instead of our own
bespoke, artisan software engineering or expensive verification.