memory: implement bump allocator
the bump allocator is dead simple - it just allocates the next available
memory, freed memory be damned. this is obviously not very good, but it
is useful for working out the kinks of writing a memory allocator in the
first place and making sure everything works.

notably, this implementation doesn't work. it will successfully compile,
but the memory the heap is using isn't actually mapped to physical
memory, so if something tries to allocate, it page faults and boot
loops. whoops!