interrupts: handle breakpoints and double faults
this commit also adds a module that sets up a gdt in rust and replaces
the gdt we wrote in assembly with it. this gdt allows us to define extra
stacks for different tasks, in particular, stacks for cpu exception
handling.

the double fault handler uses interrupt stack 0 so that it still works
if the kernel stack is overflowed. with this basic double fault handler,
triple faults should (theoretically) not be possible.