memory: unexport remap_the_kernel function
when I was originally working through the memory portion of the kernel,
all the memory initialization was performed in the kernel main function.
now that we have an init function here that calls this function, we want
to tamp down access as much as possible, since this interface
(currently, soon to be fixed) allows for users to create multiple
references to the ActivePageTable, an operation that is normally unsafe,
even though it's wrapped in the guise of a safe function. EEK!