well that didn't work
okay so obviously kill this commit if I ever come back to this.
essentially the problem is that calling temporary_table_map borrows a
bunch of things. this is fine in the simple case, but in the more
complex one of `with`, the things it is borrowing are also required in
the closure. I tried to work around that problem with allocator by
passing it through again, since we aren't using it at that point in
temporary_map, but I can't do the same thing with `self`, so now I guess
I'm stuck with the struct.