asm: run rust at high virtual memory
eventually I want to have my kernel mapped to the 510th pml4 entry. this
actually does that to a significant amount, essentially mapping the
kernel to both the 510th pml4 entry and the 0th, then calling rust at
the high virtual address.

right now there is an issue with the multiboot2 crate doing some kind of
subtraction that overflows. I actually feel like I want to write a
bootloader myself now instead of using multiboot/grub, so I can be in
control of as much as possible from startup onward. now that I have a
kernel to boot for bootloader testing, I think I can jump right in.