firmware: add ovmf firmware
ovmf has been consistently a giant pain in the ass, but it finally works
consistently if I just use this specific file as the bios. to prevent
any headaches in the future, I'm including it in the repo and using that
one explicitly instead of relying on the system installation for
anything.