make: switch back to qemu
so qemu is nicer in that the runtime definition is all defined in one
command line call, vs having to create a vm somewhere and just relying
on that to be configured correctly. originally I switched off of it
because I couldn't get qemu and ovmf to automatically load the efi
application, and I wasn't about to run the efi application manually in
the efi shell every time I wanted to test it. on a whim, I tried to boot
with my now definitely confirmed working efi application that I can
consistently build, and out of nowhere it did exactly what I wanted.

now I've hopefully frozen that part of the build system in amber. I
would really like to not have to mess with it anymore.