boot: move efi bootloader to separate module
eventually the efi bootloader will call the rest of the kernel with
everything setup the way that I want it. right now it just says hello
world.

Before I can use the new efi bootloader, I need to figure out a way to
remap the kernel to high virtual memory and run it, hopefully entirely
from rust. We will see if that is actually possible...