*: hello world with new uefi library
we are moving from the rust-uefi library, which is basically not
maintained at this point, to the uefi-rs library. Unfortunately, the
uefi-rs library is not on crates.io, so we refer to a checked out
version of it with a path. this will be useful anyway, since I'm not
convinced yet that this library fully implements the spec in every way
that I need (it's not their fault, the uefi spec is massive) so I might
have to add some things. We will see.

Either way, this build/run process is wayyyy simpler than the old one,
so I'm pretty happy so far.