boot: load kernel and call entry point
the efi bootloader now successfully calls the kernel entry point
function, confirmed by stepping through execution with gdb.

there are a lot of problems with this current implementation. right off
the bat, the kernel can't actually print anything. I suspect that it's
because the vga hardware isn't set up the way it expects. My next step
will be to stop trying to use vga to print and switch us to a serial
console.

also, the kernel isn't currently getting any of the information it needs
from the bootloader, like the current memory map that we pulled from the
firmware.

the kernel entry point function is now called `_start`. don't change it,
it's semantically important. the only way I could get the linker to
populate the `e_entry` field in the elf header was to call it that. it
could warrant more exploration in the future. it would be really cool if
instead of having to `extern "C"` this function, we could just call it
with the rust calling convention, since both sides are rust.

eh_personality isn't actually required if we use panic = abort.

there was an issue with finding the panic_impl function. unfortunately I
lost the rust-lang issue that suggested this workaround and tracks a
possible fix. oh well.

the kernel loading function isn't the cleanest implementation in it's
current state, it'll have to get cleaned up at some point.