makefile: remove vga device, add a serial port
the serial port outputs to stdio, and we also use -nographic, which
removes the issue where information would scroll offscreen with the vga
driver. uefi seems to gracefully handle the move to serial, but my
kernel did not.

it turns out that I was basically calling a random place in memory, and
coincidentally when I had a vga device, it didn't double-fault, but when
I have a serial device, it does. next I need to dig into linker scripts
and the rust target file so I can actually build a binary that looks
like what I expect. woo.