boot: talk a big game
in thinking about how to deal with remapping the kernel, I ended up
writing a dissertation in the comments. I also add memory as a
dependency, create the remap_kernel function, and print the total number
of pages in the memory map, left over from when I was trying to
understand the changes that need to be made.